% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsFilteringWidget.R
\name{ProteomicsFilteringWidget}
\alias{ProteomicsFilteringWidget}
\alias{printf}
\title{ProteomicsFilteringWidget}
\usage{
printf(...)
}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing

matrix
}
\description{
an R6 class providing clean access to a simple html message box
}
\details{
R6 Class representing a simple text display widget
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProteomicsFilteringWidget-new}{\code{ProteomicsFilteringWidget$new()}}
\item \href{#method-ProteomicsFilteringWidget-plotProteins}{\code{ProteomicsFilteringWidget$plotProteins()}}
\item \href{#method-ProteomicsFilteringWidget-maxOfVectors}{\code{ProteomicsFilteringWidget$maxOfVectors()}}
\item \href{#method-ProteomicsFilteringWidget-ui}{\code{ProteomicsFilteringWidget$ui()}}
\item \href{#method-ProteomicsFilteringWidget-server}{\code{ProteomicsFilteringWidget$server()}}
\item \href{#method-ProteomicsFilteringWidget-findCorrelations}{\code{ProteomicsFilteringWidget$findCorrelations()}}
\item \href{#method-ProteomicsFilteringWidget-clone}{\code{ProteomicsFilteringWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-new}{}}}
\subsection{Method \code{new()}}{
Create a new ProteomicsFilteringWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$new(id, tbl.data, tbl.complexes, title = "foo")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{tbl.data}}{data.frame, idiosyncratic format for DIA data}

\item{\code{tbl.complexes}}{data.frame, associates proteins (by gene name) from tbl.data into complexes}

\item{\code{title}}{character}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ProteomicsFilteringWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-plotProteins"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-plotProteins}{}}}
\subsection{Method \code{plotProteins()}}{
draw a protein vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$plotProteins(input, output, mtx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{the shiny object, access to the ui}

\item{\code{output}}{the shiny object, for drawing}

\item{\code{mtx}}{extracted from tbl.all, just the Dn timepoint columns, just the selected}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-maxOfVectors"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-maxOfVectors}{}}}
\subsection{Method \code{maxOfVectors()}}{
find the maximum of several vectors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$maxOfVectors(vectorList)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vectorList}}{list of numeric vectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
numeric
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-server"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-findCorrelations"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-findCorrelations}{}}}
\subsection{Method \code{findCorrelations()}}{
finds proteins/genes matching correlation criteria
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$findCorrelations(
  target.gene,
  target.fraction,
  threshold,
  direction,
  method,
  excluded.timepoints,
  tbl.current
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target.gene}}{character}

\item{\code{target.fraction}}{character}

\item{\code{threshold}}{numeric}

\item{\code{direction}}{character}

\item{\code{method}}{character}

\item{\code{excluded.timepoints}}{vector}

\item{\code{tbl.current}}{data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
