% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsFilteringWidget.R
\name{ProteomicsFilteringWidget}
\alias{ProteomicsFilteringWidget}
\alias{printf}
\title{ProteomicsFilteringWidget}
\usage{
printf(...)
}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing

nothing

character
}
\description{
an R6 class providing clean access to a simple html message box
}
\details{
R6 Class representing a simple text display widget
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProteomicsFilteringWidget-new}{\code{ProteomicsFilteringWidget$new()}}
\item \href{#method-ProteomicsFilteringWidget-ui}{\code{ProteomicsFilteringWidget$ui()}}
\item \href{#method-ProteomicsFilteringWidget-server}{\code{ProteomicsFilteringWidget$server()}}
\item \href{#method-ProteomicsFilteringWidget-setText}{\code{ProteomicsFilteringWidget$setText()}}
\item \href{#method-ProteomicsFilteringWidget-getText}{\code{ProteomicsFilteringWidget$getText()}}
\item \href{#method-ProteomicsFilteringWidget-clone}{\code{ProteomicsFilteringWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-new"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-new}{}}}
\subsection{Method \code{new()}}{
Create a new ProteomicsFilteringWidget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$new(
  id,
  title,
  boxWidth = 200,
  boxHeight = 30,
  fontSize = 20,
  fontColor = "black",
  backgroundColor = "beige"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{title}}{character}

\item{\code{boxWidth}}{integer, 200 by default}

\item{\code{boxHeight}}{integer, 30 by default}

\item{\code{fontSize}}{integer, 20 by default}

\item{\code{fontColor}}{character, standard CSS rgb or name}

\item{\code{backgroundColor}}{character, standard CSS rgb or name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ProteomicsFilteringWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-server"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-setText"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-setText}{}}}
\subsection{Method \code{setText()}}{
registers the specified text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$setText(newText)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newText}}{character}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-getText"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-getText}{}}}
\subsection{Method \code{getText()}}{
retrieves current text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$getText()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProteomicsFilteringWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProteomicsFilteringWidget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProteomicsFilteringWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
